requirements.txt
transformers~=4.25.1
accelerate
datasets
huggingface_hub

wikiextractor==0.1
namu-wiki-extractor==0.2.1

numpy~=1.24.1
scikit-learn~=1.1.2
scipy~=1.9.3
matplotlib~=3.6.1
pandas~=1.5.1

sacremoses==0.0.53
mosestokenizer~=1.2.1
sentencepiece==0.1.91
konlpy~=0.6.0
JPype1
openpyxl
tensorboard
tensorboardX

nltk
hgtk
jamotools
g2pk


tqdm~=4.64.1
attrdict

setuptools==59.5.0

jupyter
IPython
bertviz
einops

